+++ src/package.h
@@ -34,6 +34,15 @@ typedef struct {
   list_t *src;
 } clib_package_t;
 
+typedef struct
+{
+    int skip_cache;
+} clib_package_opts_t;
+
+
+void
+clib_package_set_opts(clib_package_opts_t opts);
+
 clib_package_t *
 clib_package_new(const char *, int);